+++ src/components/exampleCard/example-card.module.css
@@ -2,33 +2,25 @@
   max-width: 1100px;
   display: flex;
   flex-direction: column;
-  gap: 1.5rem;
+  justify-content: space-between;
 }
 
 .dropdownTest {
   color: white;
 }
 
-.text {
-  max-width: 30rem;
-  display: grid;
-  gap: 1rem;
-}
-
 .img {
   min-width: 450px;
+  width: 550px;
 }
 
 @media screen and (min-width: 1200px) {
   .card {
     flex-direction: row;
+    width: auto;
   }
 
   .left {
     flex-direction: row-reverse;
-  }
-
-  .text {
-    margin-right: 1.5rem;
   }
 }
