+++ src/components/text.module.css
@@ -2,6 +2,7 @@
   color: white;
   line-height: 2rem;
   letter-spacing: 0.05rem;
+  margin-bottom: 1rem;
 }
 
 .sm {
