+++ src/pages/index.jsx
@@ -10,7 +10,7 @@ import * as styles from "./styles/index.module.css";
 import Button from "../components/button";
 import Heading from "../components/heading";
 import Text from "../components/text";
-import ExampleCard from "../components/example-card";
+import ExampleCard from "../components/exampleCard/ExampleCard";
 
 import watchDemoIcon from "../images/watch-demo-icon.png";
 import githubIcon from "../images/github-transparent.png";
@@ -20,6 +20,7 @@ import fullClassGif from "../gifs/java/Full-Class.gif";
 import commentsGif from "../gifs/java/Comments.png";
 import mocksGif from "../gifs/java/Mocks.gif";
 import falsePositiveGif from "../gifs/java/False-Positive.gif";
+import TextContent from "../components/exampleCard/textContent/TextContent";
 
 const MainPage = ({ location }) => {
   const { t, i18n } = useTranslation();
@@ -44,12 +45,7 @@ const MainPage = ({ location }) => {
             to="https://youtu.be/p6GtTqc599Q"
             className={styles.howWorksLink}
           >
-            <img
-              src={watchDemoIcon}
-              width="32"
-              height="30"
-              alt="Video icon"
-            />
+            <img src={watchDemoIcon} width="32" height="30" alt="Video icon" />
             {t("javaHome.watchHowWorksText")}
           </Link>
 
@@ -124,12 +120,12 @@ const MainPage = ({ location }) => {
           gifAlt="Human readable tests example"
           gifPlacement="right"
         />
-
-        <ExampleCard
+        <TextContent
           heading={t("javaHome.capableForYourNeedText")}
           text={t("javaHome.capableForEveryText")}
           linkTo="/docs/java/check-requirements"
           linkText={t("javaHome.supportedTechnologiesText")}
+          containerClassName="onlyText"
         />
       </Container>
     </Layout>
