+++ src/pages/styles/index.module.css
@@ -22,6 +22,8 @@
   gap: 1.5rem;
   min-height: calc(100vh - var(--header-desktop-height));
   justify-content: center;
+
+  z-index: 1;
 }
 
 @media screen and (max-width: 991px) {
@@ -65,7 +67,8 @@
 }
 
 .examples {
-  padding-top: 0.5rem;
+  padding-top: 1.5rem;
+  padding-bottom: 3rem;
   display: grid;
   gap: 6rem;
   place-items: center;
@@ -75,4 +78,17 @@
   .examples {
     gap: 12rem;
   }
+}
+
+.gradient {
+  width: 100vw;
+  height: 103vh;
+  position: absolute;
+  top: 0;
+  background: radial-gradient(
+    circle,
+    rgba(43, 64, 40, 1) 0%,
+    rgba(3, 33, 3, 1) 40%,
+    rgba(1, 14, 1, 1) 100%
+  );
 }
